+++ headers/functions.h
@@ -1,10 +1,5 @@
-#ifndef FUNCTIONS_H
-#define FUNCTIONS_H
-
 #include "structs.h"
 
 char *pathToFile(char *str);
 bool rocketBorderCheck(Rocket *r);
 void soldierBorderCheck(Soldier *s);
-
-#endif
